reroute files, add more documentation across issue
