+++ Assets/Scripts/Rooms.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7865d00f52aa3ad4389edd1ced2c98ba
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
