+++ Assets/Scripts/Rooms/ITeleporter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6139e15aded5bad469590e6d5bdf0e77
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
