+++ Assets/Scripts/Rooms/Room.cs
@@ -2,22 +2,32 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
+/// <summary>
+/// A playable room instance.
+/// </summary>
 public class Room : MonoBehaviour
 {
-    public string Id = "";
-    public Dictionary<string, TransitionEndpoint> roomEndpoints = new Dictionary<string, TransitionEndpoint>();
+    [SerializeField] private string _id;
 
+    // Read-only accessor
+    public string Id { get => _id; }
+    private readonly Dictionary<string, TransitionEndpoint> _roomEndpoints = new Dictionary<string, TransitionEndpoint>();
 
+    /// <summary>
+    /// Prepare this room for being played in.
+    /// </summary>
     public void LoadRoom()
     {
         gameObject.SetActive(true);
     }
 
+    /// <summary>
+    /// Remove the room from play and set to inactive.
+    /// </summary>
     public void UnloadRoom()
     {
         gameObject.SetActive(false);
     }
-
 
     private void Awake()
     {
@@ -30,7 +40,7 @@ public class Room : MonoBehaviour
 
         foreach (TransitionEndpoint endPoint in endpoints)
         {
-            roomEndpoints.Add(endPoint.Id, endPoint);
+            _roomEndpoints.Add(endPoint.Id, endPoint);
         }
     }
 }
