+++ Assets/Scripts/Rooms/TransitionEndpoint.cs.meta
